metrics: heapster unable to handle sandbox continer when using containerd as container runtime.
Version v1.5.4 used in 5.5.x gravity branch does not support containerd
(basically, CRI-specific) cAdvisor metrics due to a slightly different
set of labels (specifically, the sandbox container is labelled using a
different label than the one in the Docker case). Due to this
discrepancy, heapster is unable to aggregate Pod-level metrics (which
includes network metrics) resulting in 'No Data' in Grafana.

To fix this, add support for detecting the sandbox container in CRI case
so the aggregation works as expected.